prepare migration to python 3: use absolute import instead of implicit relative import (incl: move command_line.py to __main__.py to allow module CLI call)
